requirements.txt
Flask>=1.0.3
gevent>=1.2.2
Brotli>=1.0.7
PySocks>=1.6.8
urllib3>=2.0.0
defusedxml>=0.5.0
cachetools>=4.0.0
stem>=1.8.0
Pillow>=9.0.0
